event-broker/requirements.txt
ably==2.0.12
PyYAML==6.0.2
pydantic==2.11.7
aiohttp==3.12.15
ib-async==2.0.1
tzdata==2025.2